requirements.txt
pymupdf==1.23.8
tiktoken==0.5.2
sentence-transformers==2.7.0
faiss-cpu==1.7.4
openai==1.14.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.29
pyyaml==6.0.1
python-dotenv==1.0.1
streamlit==1.32.0
numpy==1.26.4
pandas==2.2.1
tqdm==4.66.2